Initial checkin, version 0.0.1.
We've had this running in Galion for a couple of weeks now, and I
just finished turning most of the hardcoded things that other sites
will want to configure, into variables in the misc_variables table,
and most of the rest into things that are configurable in one of
the custom .pl files (binpath.pl, dbconfig.pl, mailconfig.pl;
samples of these files are provided, but the files themselves are
not checked into the repo; thus, if you only update these three .pl
files themselves and not the samples, updates will not conflict
with your edits).  I also added a config variable for the names
of the stylesheets, in case you want to do custom styling; you can
make copies of the stock stylesheets with different names, configure
it to use your copies, and then edit to your heart's content.

There aren't really setup instructions yet, or a config script, or
other such niceties, but the software is basically working as near
as I can tell, though only lightly tested as yet.  We'll be using
it for most of our wireless printing needs in Galion going forward,
so testing _will_ be happening over the coming months.

But I figure this is good enough to push out now, and then future
enhancements and bug fixes can come in future commits.